pyproject.toml
[tool.poetry]
name = "digitalocean-firewalls-ip-changer"
version = "0.1.0"
description = "IP changer for the DigitalOcean firewall rules"
authors = ["HitLuca <luca.simonetto.94@gmail.com>"]
readme="README.rst"
license="MIT"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"
requests = "^2.23.0"
ipy = "^1.00"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
vulture = "^1.3"
black = "^19.10b0"
pylint = "^2.5.0"
mypy = "^0.770"
pygments = "^2.6.1"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
update_digitalocean_firewall = "digitalocean_firewalls_ip_changer.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

